new features, CUDA support, Interface iteration

+++ runner.sh
@@ -1,10 +1,27 @@
+# runner.sh
+if [ -z ${MODEL_ROOT} ]; then 
+  echo set MODEL_ROOT
+  exit 1
+fi
+if [ -z ${CUDA_CKPT_PATH} ]; then
+  echo set CUDA_CKPT_PATH
+  exit 1
+fi
+
+if [[ "$(uname -m)" == "x86_64" ]]; then
+  export EXTRA_ARGS='--env-file env.cuda --env-file env.lcm --name cuda-lcm-sd-ui darkbit1001/cuda-lcm-sd-ui:latest'
+else 
+  export EXTRA_ARGS='---env-file env.lcm --name rknn-lcm-sd-ui darkbit1001/rknn-lcm-sd-ui:latest'
+fi
+
+
+
+set -x
 docker run --rm -it \
-  --name lcm-sr-ui \
+  --gpus all \
   --network host \
   --privileged \
-  -e PORT=4200 \
-  -e NUM_WORKERS=1 \
-  -e QUEUE_MAX=8 \
-  -e MODEL_ROOT=/models \
-  -v "$PWD/model:/models:ro,Z" \
-  lcm-sr-ui:latest
+  -e CUDA_CKPT_PATH=${CUDA_CKPT_PATH} \
+  $@ \
+  -v "${MODEL_ROOT}:/models:ro,Z" ${EXTRA_ARGS} bash  
+set +x
